requirements.txt
django
djangorestframework
factory_boy
django-timezone-field